refactor(timer_screen,route): rename timer_page to timer_screen for consistency
